% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidy_rates.R
\docType{package}
\name{aneuploidy_rates}
\alias{aneuploidy_rates}
\title{aneuploidy_rates: A package for computing the meiotic and mitotic error rates
of human embryos.}
\description{
Run find_rates to generate the error rate data.
}
\section{This package includes five main functions}{

create_embryo, prob_to_prop, take_biopsy, summarize_biopsy, and find_rates

summarize_biopsy uses create_embryo, prob_to_prop, and take_biopsy.

find_rates uses summarize_biopsy.
}

