% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidyRates-package.R
\docType{package}
\name{aneuploidyRates-package}
\alias{aneuploidyRates}
\alias{aneuploidyRates-package}
\title{aneuploidyRates: Simulates Embryonic Development to Investigate Aneuploidy Rates}
\description{
Simulate early embryo development under a range of meiotic and mitotic error rates. Compare resulting biopsies to clinically detected aneuploidy rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mccoy-lab/aneuploidyRates}
  \item Report bugs at \url{https://github.com/mccoy-lab/aneuploidyRates/issues}
}

}
\author{
\strong{Maintainer}: Angela Yang \email{qyang40@jhu.edu}

Authors:
\itemize{
  \item Sara Carioscia \email{saracarioscia@gmail.com}
}

}
\keyword{internal}
