% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rates.R
\name{find_rates}
\alias{find_rates}
\title{Return a data frame of the selected error probabilities and dispersal}
\usage{
find_rates(
  meio.range = list(0, 1),
  mito.range = list(0, 1),
  disp.range = list(0, 1),
  expected = c(0.388, 0.186, 0.426),
  tolerance = 0.05,
  num.trials = 100,
  hide.param = TRUE
)
}
\arguments{
\item{meio.range}{a double for the uniform distribution range to generate a meiotic error rate}

\item{mito.range}{a double for the uniform distribution range to generate a mitotic error rate}

\item{disp.range}{a double for the uniform distribution range to generate the extent of dispersal}

\item{expected}{a list of ratios derived from published data, used for selecting
the fitting error rates and dispersal}

\item{tolerance}{the percent of simulations to be kept near the expected values}

\item{num.trials}{the number of trials to run the simulation. Each trial}

\item{hide.param}{a boolean to show/hide the constant default parameters: num.cells,
num.chr, concordance}
}
\value{
a data frame of the corresponding embryo (prop.aneu), the selected
error rate pair, its dispersal, and biopsy information.
}
\description{
This function runs through a range of meiotic and mitotic probabilities and dispersal,
and select the most fitting values that deduces the expected data. The selection
process will be done by EasyABC.

Data for comparison:
Viotti et al,. 2021 (https://doi.org/10.1016/j.fertnstert.2020.11.041).
We leveraged their summary statistics presented in Figure 1A and calculated
a weighted average to determine percentage of euploid, mosaic, and aneuploid
biopsies across 5 clinics.

A total of 73218 embryos are collected, of which 38.8\% (28,431) are euploid,
18.6\% (13,602) are mosaic, and 42.6\% (31,185) are aneuploid.
}
\examples{
find_rates(num.trials = 40)
find_rates(
  meio.range = list(0.3, 0.5),
  mito.range = list(0.05, 0.15),
  disp.range = list (0, 0.5),
  expected = c(0.3, 0.1, 0.6),
  tolerance = 0.1,
  num.trials = 50,
  hide.param = FALSE)
}
