% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embryo.R
\name{length,Embryo-method}
\alias{length,Embryo-method}
\title{Add a method override for existing generic function length to get the number
of cells}
\usage{
\S4method{length}{Embryo}(x)
}
\arguments{
\item{x}{an Embryo object}
}
\description{
Add a method override for existing generic function length to get the number
of cells
}
