% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_prop.R
\name{mito_aneu_cells}
\alias{mito_aneu_cells}
\title{Simulate cell division and count the number of affected (aneuploid) cells.
Recursive.
The derived cells of an aneuploid cell will all be aneuploids. Therefore,
we can tally the number of affected cells mathematically.}
\usage{
mito_aneu_cells(
  cells.affected = 0,
  n.division = 8,
  prob.affected = 0.5,
  currently.affected = F
)
}
\arguments{
\item{n.division}{number of divisions simulated}

\item{prob.affected}{the likelihood of a cell to be an aneuploid.}

\item{currently.affected}{True if the current cell is an aneuploid and False otherwise}

\item{cell.affected}{current total number of affected cells}
}
\value{
total number of affected cells in this simulation
}
\description{
Simulate cell division and count the number of affected (aneuploid) cells.
Recursive.
The derived cells of an aneuploid cell will all be aneuploids. Therefore,
we can tally the number of affected cells mathematically.
}
