% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embryo.R
\name{show,Embryo-method}
\alias{show,Embryo-method}
\title{Override show function for an Embryo object}
\usage{
\S4method{show}{Embryo}(object)
}
\arguments{
\item{object}{an Embryo object}
}
\description{
Override show function for an Embryo object
}
