% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_biopsy.R
\name{take_biopsy}
\alias{take_biopsy}
\title{Takes a biopsy of an embryo object.}
\usage{
take_biopsy(em, biop.size = 5)
}
\arguments{
\item{em}{the embryo for the biopsy sampling}

\item{biop.size}{the number of cells to be sampled}
}
\value{
an integer representing the biopsy type
}
\description{
Return an int corresponding to the biopsy types:
2 = aneuploid
1 = mosaic
0 = euploid
}
\examples{
take_biopsy(tessera::Embryo(
n.cells = 100,
n.chrs = 1,
prop.aneuploid = 0.5,
dispersal = 0), biop.size =5)

}
