% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneuploidyRates-package.R
\docType{package}
\name{aneuploidy_rates}
\alias{aneuploidyRates}
\alias{aneuploidyRates-package}
\alias{aneuploidy_rates}
\title{aneuploidy_rates: A package for computing the meiotic and mitotic error rates
of human embryos.}
\description{
Run find_rates to generate the error rate data.
}
\section{This package includes five main functions}{

create_embryo, prob_to_prop, take_biopsy, summarize_biopsy, and find_rates

summarize_biopsy uses create_embryo, prob_to_prop, and take_biopsy.

find_rates uses summarize_biopsy.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mccoy-lab/aneuploidyRates}
  \item Report bugs at \url{https://github.com/mccoy-lab/aneuploidyRates/issues}
}

}
\author{
\strong{Maintainer}: Angela Yang \email{qyang40@jhu.edu}

Authors:
\itemize{
  \item Sara Carioscia \email{saracarioscia@gmail.com}
}

}
\keyword{internal}
