% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embryo.R
\name{create_embryo}
\alias{create_embryo}
\title{Create an embryo with the given proportion of aneuploidies.}
\usage{
create_embryo(
  n.cells = 200,
  n.chrs = 1,
  prop.aneuploid = 0.2,
  dispersal = 0,
  concordance = 0,
  euploidy = 2,
  rng.seed = NULL
)
}
\arguments{
\item{n.cells}{the number of cells in the embryo}

\item{n.chrs}{the number of chromosome pairs per cell}

\item{prop.aneuploid}{the proportion vector of aneuploid cells (0-1) per chromosome}

\item{dispersal}{the dispersion vector of the aneuploid cells (0-1)}

\item{concordance}{the concordance between aneuploid cells for each chromosome (0-1).}

\item{euploidy}{the number of copies of a chromosome to consider euploid. For a diploid embryo this should be 2.}

\item{rng.seed}{the seed for the rRandom Number Generation (RNG). Defaults to NULL. Use this to get the same embryo each time}
}
\value{
an Embryo object
}
\description{
Create an embryo with the given proportion of aneuploidies.
}
\examples{
embryo <- create_embryo(n.cells = 200, n.chrs = 1,  prop.aneuploid = 0.2,
dispersal =  0.9)

embryo <- create_embryo(n.cells = 200, n.chrs = 1,  prop.aneuploid = 0.2,
dispersal =  0.9, rng.seed = 42)

embryo <- create_embryo(n.cells = 200, n.chrs = 3,  prop.aneuploid = 0.2,
dispersal =  0.9, concordance = 1)

embryo <- create_embryo(n.cells = 250, n.chrs = 3,
prop.aneuploid = c(0.2, 0.1, 0.4), dispersal =  0.9)
}
