% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_embryo.R
\docType{class}
\name{Embryo-class}
\alias{Embryo-class}
\title{This file creates an embryo object.
code used from line 9-511 are mainly referenced from the Tessera package,
with more lines of comments, slight changes in line 332-343, and removal
of the graphic}
\description{
(Maybe just use the original package? Then how do we write
this file if there are no new functions?)
An S4 class to model an embryo
}
\section{Slots}{

\describe{
\item{\code{x}}{numeric. x coordinates of cells}

\item{\code{y}}{numeric.  y coordinates of cells}

\item{\code{z}}{numeric.  z coordinates of cells}

\item{\code{aneu}}{numeric. fraction of aneuploid cells}

\item{\code{disp}}{numeric. fraction of dispersal of cells}

\item{\code{dists}}{data.frame. pairwise distances between cells}

\item{\code{neighbours}}{data.frame, whether two cells are near each other}

\item{\code{euploidy}}{numeric. number of chromosomes considered euploid}

\item{\code{ploidy}}{data.frame. number of chromosomes per cell}
}}

