% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_prop.R
\name{prob_to_prop}
\alias{prob_to_prop}
\title{This file converts the probabilities of being affected
by meiotic and mitotic errors to proportions of
aneuploid cells within an embryo.
Convert the probabilities to proportion.}
\usage{
prob_to_prop(prob.meio, prob.mito, num.division = 8)
}
\arguments{
\item{prob.meio}{the probability of having a meiotic error}

\item{prob.mito}{the probability of having a mitotic error}

\item{num.division}{the total number of divisions in this embryo}
}
\value{
proportion of totally affected (aneuploidy) cells within
this embryo.
}
\description{
This file converts the probabilities of being affected
by meiotic and mitotic errors to proportions of
aneuploid cells within an embryo.
Convert the probabilities to proportion.
}
