% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_biopsy.R
\name{summarize_biopsy}
\alias{summarize_biopsy}
\title{A wrapper function for generating the biopsy results}
\usage{
summarize_biopsy(num.em = 1000, meio, mito)
}
\arguments{
\item{num.em}{the number of embryos to be created}

\item{meio}{the probability of having a meiotic error}

\item{mito}{the probability of having a meiotic error
(try with probability ranges instead of )}
}
\description{
A wrapper function for generating the biopsy results
}
