% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_biopsy.R
\name{take_biopsy}
\alias{take_biopsy}
\title{Take a biopsy of the given embryo, return an int corresponding to the biopsy types
2 = aneuploid
1 = mosaic
0 = euploid}
\usage{
take_biopsy(em, biop.size = 5)
}
\arguments{
\item{em}{the embryo for the biopsy sampling}

\item{biop.size}{the number of cells to be sampled}
}
\value{
an integer representing the biopsy type
}
\description{
Take a biopsy of the given embryo, return an int corresponding to the biopsy types
2 = aneuploid
1 = mosaic
0 = euploid
}
